create UserRepositoryImpl add changes in UnistartModule some changes in User.kt and User.sq

+++ server/src/main/kotlin/ru/spbstu/application/auth/entities/User.kt
@@ -5,7 +5,7 @@ data class User(
     val phoneNumber: PhoneNumber,
     val avatar: Avatar,
     val occupation: Occupation,
-    val availableStepsCount: Int
+    val availableStepsCount: Long
 ) {
     @JvmInline
     value class Id(val value: Long)
